feat: validator 생성, 인원 마감 조건 추가
